Bullet separete object
