+++ Diepio_Client/FinaleSignalR-Client/FinaleSignalR-Client.csproj
@@ -175,13 +175,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Controls\MapControls.cs">
+    <Compile Include="Objects\Bullet.cs" />
+    <Compile Include="Objects\MapControls.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Form1.cs" />
+    <Compile Include="Form1.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
+    <Compile Include="Objects\Player.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Web\Communication.cs" />
