+++ Diepio_Client/FinaleSignalR-Client/Objects/MapControls.cs
